x/verify/types/message_kimchi.go
package types

import (
	errorsmod "cosmossdk.io/errors"
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

var _ sdk.Msg = &MsgKimchi{}

func NewMsgKimchi(creator string, proof string) *MsgKimchi {
	return &MsgKimchi{
		Creator: creator,
		Proof:   proof,
	}
}

func (msg *MsgKimchi) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
